Refactor data class
